Refactor jump_to_label implementation
This splits the implemention of jump_to_label into more parts that allow
it to be reused in the future for other on-screen pickers.

The main way it will be used is by generating some ranges for the labels
(see generate_viewport_token_ranges, the function that generates the
ranges of visible tokens in the current viewport) and giving those labels
to the prompt_for_label function. prompt_for_label calls a callback
when a label has been typed.

An example of how this could be used in the future is to pick a selection
to set as the primary selection.