Add add_selection_on_word command
This command uses the on-screen word picker, the same as goto_word and
extend_word, but instead of removing your existing selections, it places
a new selection on the picked word. This allows you to select arbitrary
words.

By default, this is bound to gW (as opposed to goto_word, which is 'gw').